rename ppo_actor.py to train_actor.py
